Added experiment scripts
Added experiment suites in justfile

Split genesis-block creation from creds build to avoid redundant steps in deployments between different experiments.

+++ ordererbuilder/init.sh
@@ -5,15 +5,26 @@ txgen_config=$2
 out=$3
 channel_id=$4
 
-"$BINS_PATH"/cryptogen generate \
+echo "Parameters passed: '$crypto_config', '$txgen_config', '$out', '$channel_id'"
+
+if [ -n "$crypto_config" ]; then \
+  echo "Executing cryptogen..."
+  "$BINS_PATH"/cryptogen generate \
     --config="$crypto_config" \
     --output="$out/orgs"
+else \
+  echo "Skipping cryptogen..."
+fi
 
-cp "$txgen_config" /usr/local/configtx.yaml
+if [ -n "$txgen_config" ] || [ -n "$channel_id" ]; then \
+  echo "Executing blockgen creation..."
+  cp "$txgen_config" /usr/local/configtx.yaml
+  "$BINS_PATH"/configtxgen \
+      -outputBlock "$out/genesisblock" \
+      -profile SampleDevModeEtcdRaft \
+      -channelID "$channel_id"
+  rm /usr/local/configtx.yaml
+else \
+  echo "Skipping blockgen creation..."
+fi
 
-"$BINS_PATH"/configtxgen \
-    -outputBlock "$out/genesisblock" \
-    -profile SampleDevModeEtcdRaft \
-    -channelID "$channel_id"
-
-rm /usr/local/configtx.yaml
